refactor: Extracted resolver into different classes
